Move `AnalysisBuffer` to analysis package
