+++ packages/engine/src/output/buffer/mod.rs
@@ -1,19 +1,10 @@
-use std::{collections::HashMap, sync::Arc};
-
 mod part;
 
-use execution::package::simulation::{
-    output::{
-        analysis::{AnalysisOutput, AnalysisOutputConfig, AnalysisSingleOutput},
-        OutputPackageName,
-    },
-    OutputPackagesSimConfig, PackageName,
-};
-use serde::Serialize;
+use execution::package::simulation::{output::analysis::AnalysisBuffer, OutputPackagesSimConfig};
 use simulation_structure::{ExperimentId, SimulationShortId};
 
 pub use self::part::{remove_experiment_parts, OutputPartBuffer};
-use crate::output::error::{Error, Result};
+use crate::output::error::Result;
 
 // TODO: We might want to use a temporary folder (like "/tmp" or "/var/tmp") instead.
 const RELATIVE_PARTS_FOLDER: &str = "./parts";
@@ -33,41 +24,6 @@ impl Buffers {
             // TODO: This should be dynamically created by the output packages
             json_state: OutputPartBuffer::new("json_state", exp_id, sim_id)?,
             analysis: AnalysisBuffer::new(output_packages_sim_config)?,
-        })
-    }
-}
-
-// TODO: These should live in the respective output package really
-#[derive(Serialize)]
-pub struct AnalysisBuffer {
-    pub manifest: String,
-    pub buffers: HashMap<Arc<String>, Vec<AnalysisSingleOutput>>,
-}
-
-impl AnalysisBuffer {
-    pub fn new(output_packages_config: &OutputPackagesSimConfig) -> Result<AnalysisBuffer> {
-        let value = output_packages_config
-            .map
-            .get(&PackageName::Output(OutputPackageName::Analysis))
-            .ok_or_else(|| Error::from("Missing analysis config"))?;
-        let config: AnalysisOutputConfig = serde_json::from_value(value.clone())?;
-        let buffer = AnalysisBuffer {
-            manifest: config.manifest.clone(),
-            buffers: config.outputs.keys().map(|v| (v.clone(), vec![])).collect(),
-        };
-        Ok(buffer)
-    }
-
-    pub fn add(&mut self, output: AnalysisOutput) -> Result<()> {
-        output.inner.into_iter().try_for_each(|(name, output)| {
-            self.buffers
-                .get_mut(&name)
-                .ok_or_else(|| {
-                    Error::from(format!("Missing output buffer when persisting: {}", &name))
-                })?
-                .push(output);
-
-            Ok(())
         })
     }
 }
