+++ packages/engine/src/output/error.rs
@@ -6,6 +6,9 @@ pub enum Error {
     #[error("Output error: {0}")]
     Unique(String),
 
+    #[error("Execution error: {0}")]
+    Execution(#[from] execution::Error),
+
     #[error("Deserialization error: {0}")]
     FromSerde(#[from] serde_json::Error),
